Add support for RequestBlocklist feature (#3275)
The RequestBlocklist feature allows for arbitrary requests to be blocked. Being
able to do that will be useful for resolving some website breakage issues, where
the necessary fix is to block a request (rather than to allow one).

This change includes MV2 and MV3 implementations of the feature, unit tests and
integration tests. See the reference tests:
https://github.com/duckduckgo/privacy-reference-tests/tree/main/request-blocklist